requirements.txt
absl-py==2.0.0
accelerate==0.26.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.14.0
beartype==0.20.2
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli==1.0.9
bypy==1.8.5
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
crcmod==1.7
cryptography==41.0.7
cycler==0.12.1
Cython==3.0.12
datasets==2.15.0
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.25.0
dill==0.3.7
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
hear21passt==0.0.26
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib-resources==6.1.0
ipykernel==6.28.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
lmdb==0.9.31
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.2.0
modelscope==1.10.0
moviepy==1.0.3
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
oss2==2.18.3
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
Pillow==23.2.1
platformdirs==3.10.0
pooch==1.7.0
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydub==0.25.1
pyg-lib==0.3.1+pt113cu116
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-telegram-bot==21.10
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resampy==0.4.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
sacremoses==0.1.1
safetensors==0.5.3
scikit-fuzzy==0.5.0
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.11.2
Send2Trash==1.8.2
sentence-transformers==3.4.1
Shapely==1.8.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
submitit==1.5.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.18.0
tf-slim==1.1.0
threadpoolctl==3.2.0
tifffile==2024.8.30
tiktoken==0.5.2
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.0
torch-cluster==1.6.1+pt113cu116
torch-scatter==2.1.1+pt113cu116
torch-sparse==0.6.17+pt113cu116
torch-spline-conv==1.2.2+pt113cu116
torch_geometric==2.4.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.26.1
transformers-stream-generator==0.0.4
typeguard==2.13.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
